Moved TestBatchPoints
